refactor: Reverted handle_message abstractions
