+++ crates/libtortillas/src/peers/transport_messages.rs
@@ -1,13 +1,6 @@
-use crate::errors::PeerTransportError;
-
-use super::{Peer, PeerKey};
+use super::Peer;
 
 #[derive(Debug, Clone)]
 pub enum TransportCommand {
    Connect { peer: Peer },
-}
-
-#[derive(Debug)]
-pub enum TransportOutput {
-   Handshake(Result<PeerKey, PeerTransportError>),
 }
